requirements.txt
streamlit==1.44.1
transformers==4.49.0
torch==2.6.0
Pillow==11.1.0
